+++ R/ArchiveAsync.R
@@ -100,11 +100,7 @@ ArchiveAsync = R6Class(
     #' List of named lists of point values.
     #' @param xss_extra (list of named `list()` | `NULL`)\cr
     #' List of named lists of additional information.
-    #' @param extra (list of named `list()` | `NULL`)\cr
-    #' Deprecated argument for additional information.
-    #' Use `xss_extra` instead.
-    push_points = function(xss, xss_extra = NULL, extra = NULL) {
-      xss_extra = xss_extra %??% extra
+    push_points = function(xss, xss_extra = NULL) {
       if (self$check_values) {
         map(xss, self$search_space$assert)
       }
@@ -124,11 +120,7 @@ ArchiveAsync = R6Class(
     #' Named list of point values.
     #' @param xs_extra (named `list()` | `NULL`)\cr
     #' Named list of additional information.
-    #' @param extra (named `list()` | `NULL`)\cr
-    #' Deprecated argument for additional information.
-    #' Use `xs_extra` instead.
-    push_point = function(xs, xs_extra = NULL, extra = NULL) {
-      xs_extra = xs_extra %??% extra
+    push_point = function(xs, xs_extra = NULL) {
       if (self$check_values) {
         self$search_space$assert(xs)
       }
@@ -143,11 +135,7 @@ ArchiveAsync = R6Class(
     #' List of named lists of point values.
     #' @param xss_extra (list of named `list()` | `NULL`)\cr
     #' List of named lists of additional information.
-    #' @param extra (list of named `list()` | `NULL`)\cr
-    #' Deprecated argument for additional information.
-    #' Use `xss_extra` instead.
-    push_running_points = function(xss, xss_extra = NULL, extra = NULL) {
-      xss_extra = xss_extra %??% extra
+    push_running_points = function(xss, xss_extra = NULL) {
       if (self$check_values) {
         map(xss, self$search_space$assert)
       }
@@ -167,11 +155,7 @@ ArchiveAsync = R6Class(
     #' Named list of point values.
     #' @param xs_extra (named `list()` | `NULL`)\cr
     #' Named list of additional information.
-    #' @param extra (named `list()` | `NULL`)\cr
-    #' Deprecated argument for additional information.
-    #' Use `xs_extra` instead.
-    push_running_point = function(xs, xs_extra = NULL, extra = NULL) {
-      xs_extra = xs_extra %??% extra
+    push_running_point = function(xs, xs_extra = NULL) {
       if (self$check_values) {
         self$search_space$assert(xs)
       }
@@ -278,11 +262,7 @@ ArchiveAsync = R6Class(
     #' List of named lists of transformed point values.
     #' @param yss_extra (list of named `list()` | `NULL`)\cr
     #' List of named lists of additional information.
-    #' @param extra (list of named `list()` | `NULL`)\cr
-    #' Deprecated argument for additional information.
-    #' Use `yss_extra` instead.
-    finish_points = function(keys, yss, x_domains, yss_extra = NULL, extra = NULL) {
-      yss_extra = yss_extra %??% extra
+    finish_points = function(keys, yss, x_domains, yss_extra = NULL) {
       timestamp_ys = Sys.time()
       yss_extra = if (is.null(yss_extra)) {
         map(x_domains, function(x_domain) list(x_domain = list(x_domain), timestamp_ys = timestamp_ys))
@@ -305,11 +285,7 @@ ArchiveAsync = R6Class(
     #' Named list of transformed point values.
     #' @param ys_extra (named `list()` | `NULL`)\cr
     #' Named list of additional information.
-    #' @param extra (named `list()` | `NULL`)\cr
-    #' Deprecated argument for additional information.
-    #' Use `ys_extra` instead.
-    finish_point = function(key, ys, x_domain, ys_extra = NULL, extra = NULL) {
-      ys_extra = ys_extra %??% extra
+    finish_point = function(key, ys, x_domain, ys_extra = NULL) {
       ys_extra = c(list(x_domain = list(x_domain), timestamp_ys = Sys.time()), ys_extra)
       self$rush$finish_tasks(key, list(ys), extra = list(ys_extra))
     },
@@ -353,7 +329,7 @@ ArchiveAsync = R6Class(
     #' Named list of additional information.
     push_result = function(key, ys, x_domain, extra = NULL) {
       .Deprecated(new = "finish_point", old = "push_result")
-      self$finish_point(key, ys, x_domain, extra = extra)
+      self$finish_point(key, ys, x_domain, ys_extra = extra)
     },
 
     #' @description